pyproject.toml
[project]
name = "bank-account-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.123.9",
    "pytest-cov>=7.0.0",
    "sqlmodel>=0.0.27",
    "strawberry-graphql[fastapi]>=0.287.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
